permissions: fix read permissions for deposits
