+++ cap/modules/deposit/search.py
@@ -26,7 +26,6 @@
 
 from elasticsearch_dsl import Q, TermsFacet
 from flask import abort  # has_request_context
-from flask import session
 from flask_login import current_user
 from invenio_search import RecordsSearch
 from invenio_search.api import DefaultFilter
@@ -46,23 +45,23 @@ def deposits_filter():
 
     Otherwise, it filters out any deposit where user is not the owner.
     """
-    # @TOFIX check request_context
-    # if not has_request_context() or admin_permission_factory(None).can():
-    #     return Q()
-
     if admin_permission_factory(None).can():
         return Q()
 
     if current_user.is_authenticated:
-        q1 = Q('match',
-               **{'_deposit.owners': getattr(current_user, 'email', None)})
-        q2 = Q('match',
-               **{'_access.deposit-read.users': getattr(current_user,
-                                                        'id', 0)})
-        q3 = Q('terms',
-               **{'_access.deposit-read.roles': session.get('roles', [])})
+        roles = [role.id for role in current_user.roles]
 
-        q = q1 | q2 | q3
+        q = Q('multi_match',
+              query=current_user.id,
+              fields=[
+                  '_access.deposit-read.users',
+                  '_access.deposit-admin.users',
+                  '_deposit.owners'
+              ]) | \
+            Q('terms',
+              **{'_access.deposit-read.roles': roles}) | \
+            Q('terms',
+              **{'_access.deposit-admin.roles': roles})
 
         return q
