+++ tests/integration/test_api.py
@@ -49,276 +49,6 @@ def test_example(app, users, auth_headers_for_user, json_headers):
         assert resp.data == 'Pong'
 
 
-######################
-# api/deposits/  [GET]
-######################
-def test_get_deposits_when_user_not_logged_in_returns_403(app, users):
-    with app.test_client() as client:
-        resp = client.get('/deposits/')
-
-        assert resp.status_code == 403
-
-
-def test_get_deposits_when_superuser_returns_all_deposits(app, users,
-                                                          auth_headers_for_superuser,
-                                                          create_deposit):
-    with app.test_client() as client:
-        deposits = [
-            create_deposit(users['cms_user'], 'cms-analysis-v1.0.0'),
-            create_deposit(users['lhcb_user'], 'lhcb-v0.2.0'),
-            create_deposit(users['alice_user'], 'alice-analysis-v1.0.1'),
-            create_deposit(users['atlas_user'], 'atlas-analysis-v0.0.3'),
-        ]
-
-        resp = client.get('/deposits/', headers=auth_headers_for_superuser)
-        hits = json.loads(resp.data)['hits']['hits']
-
-        assert resp.status_code == 200
-        assert len(hits) == len(deposits)
-
-
-def test_get_deposits_when_normal_user_returns_only_his_deposits(app, db, users,
-                                                                 auth_headers_for_user,
-                                                                 create_deposit
-                                                                 ):
-    with app.test_client() as client:
-        user_deposits_ids = [x['_deposit']['id'] for x in [
-            create_deposit(users['cms_user'], 'cms-analysis-v0.0.1'),
-            create_deposit(users['cms_user'], 'cms-questionnaire-v0.0.1'),
-        ]]
-
-        create_deposit(users['cms_user2'], 'cms-analysis-v0.0.1'),
-        create_deposit(users['lhcb_user'], 'lhcb-v0.0.1'),
-        create_deposit(users['alice_user'], 'alice-analysis-v0.0.1'),
-        headers = auth_headers_for_user(users['cms_user'])
-
-        resp = client.get('/deposits/', headers=headers)
-
-        hits = json.loads(resp.data)['hits']['hits']
-
-        assert resp.status_code == 200
-        assert len(hits) == len(user_deposits_ids)
-        for x in hits:
-            assert x['metadata']['_deposit']['id'] in user_deposits_ids
-
-
-def test_get_deposits_user_see_deposits_that_has_read_access_to(app, db, users,
-                                                                auth_headers_for_user,
-                                                                json_headers,
-                                                                create_deposit
-                                                                ):
-    user = users['cms_user']
-    other_user = users['cms_user2']
-    with app.test_client() as client:
-        deposits = [
-            create_deposit(user, 'cms-analysis-v0.0.1'),
-            create_deposit(user, 'cms-questionnaire-v0.0.1'),
-        ]
-
-        # other user cant see any deposits
-        resp = client.get('/deposits/', headers=auth_headers_for_user(other_user))
-        hits = json.loads(resp.data)['hits']['hits']
-        assert len(hits) == 0
-
-        permissions = [{
-            'email': other_user.email,
-            'type': 'user',
-            'op': 'add',
-            'action': 'deposit-read'
-        }]
-
-        for deposit in deposits:
-            resp = client.post('/deposits/{}/actions/permissions'.format(deposit['_deposit']['id']),
-                               headers=auth_headers_for_user(user) + json_headers,
-                               data=json.dumps(permissions))
-
-
-            assert resp.status_code == 201
-
-        # sometimes es nees refresh
-        current_search.flush_and_refresh('deposits')
-        resp = client.get('/deposits/', headers=auth_headers_for_user(other_user))
-        hits = json.loads(resp.data)['hits']['hits']
-        assert len(hits) == 2
-
-
-def test_get_deposits_with_basic_json_serializer_returns_correct_fields_for_all_deposits(app,
-                                                                                         users,
-                                                                                         jsonschemas_host,
-                                                                                         auth_headers_for_superuser,
-                                                                                         create_deposit):
-    deposit = create_deposit(users['superuser'], 'cms-analysis-v0.0.1', {
-        '$schema': 'https://{}/schemas/deposits/records/cms-analysis-v0.0.1.json'.format(jsonschemas_host),
-        'basic_info': {
-            'analysis_number': 'dream_team',
-            'people_info': [{}]
-        }
-    })
-    metadata = deposit.get_record_metadata()
-
-    with app.test_client() as client:
-        resp = client.get('/deposits/',
-                          headers=[('Accept', 'application/basic+json')] +
-                                  auth_headers_for_superuser)
-
-        res = json.loads(resp.data)['hits']['hits'][0]
-
-        assert resp.status_code == 200
-        assert res == {
-            'created': metadata.created.strftime('%Y-%m-%dT%H:%M:%S.%f+00:00'),
-            'metadata': {
-                '$schema': deposit['$schema'],
-                'basic_info': {
-                    'people_info': [{}],
-                    'analysis_number': 'dream_team'
-                }
-            },
-            'pid': deposit['_deposit']['id'],
-            'updated': metadata.updated.strftime('%Y-%m-%dT%H:%M:%S.%f+00:00'),
-        }
-
-###########################
-# api/deposits/{pid}  [GET]
-###########################
-def test_get_deposits_with_given_id_superuser_can_see_deposits_created_by_others(app,
-                                                                                 users,
-                                                                                 auth_headers_for_superuser,
-                                                                                 create_deposit):
-    deposit = create_deposit(users['alice_user'], 'alice-analysis-v0.0.1')
-
-    with app.test_client() as client:
-        resp = client.get('/deposits/{}'.format(deposit['_deposit']['id']),
-                          headers=auth_headers_for_superuser)
-
-        assert resp.status_code == 200
-
-
-def test_get_deposits_with_given_id_creator_can_see_his_deposit(app,
-                                                                users,
-                                                                auth_headers_for_user,
-                                                                create_deposit):
-    user = users['alice_user']
-    deposit = create_deposit(user, 'alice-analysis-v0.0.1')
-
-    with app.test_client() as client:
-        resp = client.get('/deposits/{}'.format(deposit['_deposit']['id']),
-                          headers=auth_headers_for_user(user))
-
-        assert resp.status_code == 200
-
-
-def test_get_deposits_with_given_id_other_member_of_collaboration_cant_see_deposit(app,
-                                                                                   users,
-                                                                                   auth_headers_for_user,
-                                                                                   create_deposit):
-    user, other_user = users['alice_user'], users['alice_user2'] 
-    deposit = create_deposit(user, 'alice-analysis-v0.0.1')
-
-    with app.test_client() as client:
-        resp = client.get('/deposits/{}'.format(deposit['_deposit']['id']),
-                          headers=auth_headers_for_user(other_user))
-
-        assert resp.status_code == 403
-        
-def test_get_deposits_other_user_with_read_acces_can_see_deposit(app,
-                                                                 users,
-                                                                 auth_headers_for_user,
-                                                                 json_headers,
-                                                                 create_deposit):
-    user, other_user = users['alice_user'], users['alice_user2'] 
-    deposit = create_deposit(user, 'alice-analysis-v0.0.1')
-    permissions = [{
-        'email': other_user.email,
-        'type': 'user',
-        'op': 'add',
-        'action': 'deposit-read'
-    }] 
-
-    with app.test_client() as client:
-        client.post('/deposits/{}/actions/permissions'.format(deposit['_deposit']['id']),
-                    headers=auth_headers_for_user(user) + json_headers, data=json.dumps(permissions))
-
-        resp = client.get('/deposits/{}'.format(deposit['_deposit']['id']),
-                          headers=auth_headers_for_user(other_user))
-
-        assert resp.status_code == 200
-
-
-def test_get_deposits_with_given_id_with_basic_json_serializer_returns_all_correct_fields(app,
-                                                                                          users,
-                                                                                          jsonschemas_host,
-                                                                                          auth_headers_for_superuser,
-                                                                                          create_deposit):
-    deposit = create_deposit(users['superuser'], 'cms-analysis-v0.0.1', {
-        '$schema': 'https://{}/schemas/deposits/records/cms-analysis-v0.0.1.json'.format(jsonschemas_host),
-        'basic_info': {
-            'analysis_number': 'dream_team',
-            'people_info': [{}]
-        }
-    })
-    metadata = deposit.get_record_metadata()      
-
-    with app.test_client() as client:
-        resp = client.get('/deposits/{}'.format(deposit['_deposit']['id']),
-                          headers=[('Accept', 'application/basic+json')] +
-                                  auth_headers_for_superuser)
-
-        res = json.loads(resp.data)
-
-        assert resp.status_code == 200
-        assert res == {
-            'created': metadata.created.strftime('%Y-%m-%dT%H:%M:%S.%f+00:00'),
-            'metadata': {
-                '$schema': deposit['$schema'],
-                'basic_info': {
-                    'people_info': [{}],
-                    'analysis_number': 'dream_team'
-                }
-            },
-            'pid': deposit['_deposit']['id'],
-            'updated': metadata.updated.strftime('%Y-%m-%dT%H:%M:%S.%f+00:00'),
-        }                       
-
-
-def test_get_deposits_with_given_id_with_permissions_json_serializer_returns_correct_data(app,
-                                                                                          auth_headers_for_superuser,
-                                                                                          deposit):
-    with app.test_client() as client:
-        resp = client.get('/deposits/{}'.format(deposit['_deposit']['id']),
-                          headers=[('Accept', 'application/permissions+json')] +
-                                  auth_headers_for_superuser)
-
-        res = json.loads(resp.data)
-
-        assert res == {
-            'permissions': {
-                'deposit-admin': {'roles': [], 'users': ['superuser@cern.ch']}, 
-                'deposit-read': {'roles': [], 'users': ['superuser@cern.ch']}, 
-                'deposit-update': {'roles': [], 'users': ['superuser@cern.ch']}
-            }
-        }
-
-
-def test_deposit_clone(app, db, users,
-                       auth_headers_for_user,
-                       json_headers,
-                       create_deposit):
-    deposit = create_deposit(users['cms_user'], 'cms-analysis-v0.0.1')
-    headers = auth_headers_for_user(users['cms_user']) + json_headers
-    pid = deposit['_deposit']['id']
-
-    with app.test_client() as client:
-        resp = client.post('/deposits/{}/actions/publish'.format(pid),
-                           headers=headers)
-
-        cloned_deposit = client.post('/deposits/{}/actions/clone'.format(pid),
-                                     headers=headers)
-        res = json.loads(cloned_deposit.data)
-
-        assert 201 == cloned_deposit.status_code
-        assert 'cloned_from' in res['metadata']['_deposit']
-        assert pid == res['metadata']['_deposit']['cloned_from']['value']
-
 # #######################################
 # # api/deposits/{pid}  [DELETE]
 # #######################################
